+++ EmpiriaBMS.Models/Models/Discipline.cs
@@ -9,11 +9,11 @@ public class Discipline : Entity
     public long EstimatedHours { get; set; }
 
     [Required]
-    public int ProjectId { get; set; }
+    public long ProjectId { get; set; }
     public Project? Project { get; set; }
 
     [Required]
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public DisciplineType? Type { get; set; }
 
     public ICollection<Deliverable>? Deliverables { get; set; }
