+++ EmpiriaBMS.Models/Models/Payment.cs
@@ -17,11 +17,11 @@ public class Payment : Entity
     public string? Description { get; set; }
 
     [Required]
-    public int TypeId { get; set; }
+    public long TypeId { get; set; }
     public PaymentType? Type { get; set; }
 
     [Required]
-    public int InvoiceId { get; set; }
+    public long InvoiceId { get; set; }
 
     public Invoice? Invoice { get; set; }
 }
